+++ sw.js
@@ -1,4 +1,4 @@
-var snapVersion = '10-rc2',
+var snapVersion = '10-rc3',
     cacheName = `snap-pwa-${snapVersion}`,
     filesToCache = [
         'snap.html',