+++ src/gui.js
@@ -91,7 +91,7 @@ modules.gui = '2024-July-24';
 
 // Declarations
 
-var SnapVersion = '10-rc5';
+var SnapVersion = '10-rc6';
 
 var IDE_Morph;
 var ProjectDialogMorph;