+++ HISTORY.md
@@ -124,6 +124,7 @@
 
 ### 2024-07-16
 * German translation update
+* translation support & German translation for embroidery library
 
 ### 2024-07-15
 * updated v10 change log/release notes