+++ HISTORY.md
@@ -124,6 +124,7 @@
 
 ### 2024-07-17
 * store: incremented app version to rc
+* extensions: added "www" url prefix variant to birdbrain's extension entry 
 
 ### 2024-07-16
 * German translation update