+++ HISTORY.md
@@ -126,6 +126,7 @@
 * store: incremented app version to rc
 * extensions: added "www" url prefix variant to birdbrain's extension entry
 * MQTT extension update, thanks, Simon and Xavier!
+* v10-rc2
 
 ### 2024-07-16
 * German translation update