+++ HISTORY.md
@@ -125,6 +125,7 @@
 
 ### 2024-07-22
 * more help screens, thanks, Brian and the forum gang!
+* v10-rc3
 
 ### 2024-07-18
 * gui: refresh IDE when loading a library from the cloud to support customized primitives