+++ HISTORY.md
@@ -124,6 +124,9 @@
     * German
 
 ### 2024-07-22
+* edited help screen for doDeleteBlock, converted to PNG, reduced file size
+
+### 2024-07-22
 * more help screens, thanks, Brian and the forum gang!
 * v10-rc3