+++ HISTORY.md
@@ -126,6 +126,7 @@
 
 ### 2024-07-24
 * widgets: added ability to switch octaves in the piano keyboard widget, thanks, ego-lay-atman-bay!
+* gui: added credits for ego-lay-atman-bay
 
 ### 2024-07-22
 * edited help screen for doDeleteBlock, converted to PNG, reduced file size