+++ HISTORY.md
@@ -128,6 +128,7 @@
 * widgets: added ability to switch octaves in the piano keyboard widget, thanks, ego-lay-atman-bay!
 * gui: added credits for ego-lay-atman-bay
 * v10-rc6
+* reverted piano keyboard behavior
 
 ### 2024-07-22
 * edited help screen for doDeleteBlock, converted to PNG, reduced file size